Lives-Related Functions
setCurrentLives(amount) - Arbitrarily set the # of lives to the passed-in amount

changeNumberOfLives(addOrSubtract, amount) - Dynamically add or subtract from the current lives count on the fly. 

addOrSubtract is a string, "add" or "subtract" are acceptable and each change the logic used based on which is passed in.

amount is a number and gets used to add or subtract the desired value.

updateLivesOSD()
This is called in the render() function to ensure that the game's display of the current lives is updated in real time to account for changes.