Score Related Functions
setCurrentScore(amount) - Arbitrarily set the score to the passed-in amount

changeCurrentScore(addOrSubtract, amount) - Dynamically add or subtract from the current score on the fly. 

addOrSubtract is a string, "add" or "subtract" are acceptable and each change the logic used based on which is passed in.

amount is a number and gets used to add or subtract the desired value.

updateLivesOSD()
This is called in the render() function to ensure that the game's display of the current score is updated in real time to account for changes.

Testing code was placed in the create() and stopTimer() functions to show these functions working.