Frog Movement / Sprite Changes
Disabled Frog Rotation after being unable to get it to rotate the object 'in place' (this restored normal movement behavior.

Added function:

displaySpriteDebug(spriteObject, enabled)

If we pass this function the name of a sprite object and true it will display a bunch of info related to it and also highlight the objects physics body for physics troubleshooting. This was invaluable in determining what was happening with our rotation attempts.

I also doubled the size of the Frog (within it's artwork) and increased the size of the png for animation (though the animation is not yet in place).