Goals No Longer Kill You + More Accurate Frog Collision + NickCageMode Version 0.1
In the process of trying to discover why goals kill you it was discovered it had to do with a slight overlap between the frog sprite and the first row of car sprites upon respawn after reaching the goal.

Moved the frog / obstacle collision detection away from the overlap function which was looking at sprite dimensions and to a physics signal that detects collisions the frog makes and decides whether to kill it based on what it hit.

Also included a fun boolean based on a 'mistake' that happened while re-enabling physics.