Fixed Firefox/Edge Game Rendering + Ensured Score Resets After Game Over
There was a line in the CSS that Chrome was ignoring that Firefox and Edge were not that was making the game 80% transparent in those browsers.

While testing that change I discoverd that if you save one frog and then game over/restart - your score does not go to 0 after the game over. Corrected that by adding a call to the setCurrentScore() function as part of the 'Enter after Game Over' reset logic.